add basic console based connection to the frontend
I need this to test the backend.
Two players can now start a game, but there are still a few TODOs, as well as a little design flaw. When Player a sends his name and connects to player b, the later has no chance to send his name. Therefore the name has to be send before, see the appropiate TODO